pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Mamoru Matsunaga <pineb1ue10ngtree@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
astropy = "^6.0.0"
pyswisseph = "^2.10.3.2"
pytz = "^2023.3.post1"
matplotlib = "^3.8.2"
pandas = "^2.1.3"
loguru = "^0.7.2"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
pyyaml = "^6.0.1"
pandera = "^0.17.2"
dependency-injector = "^4.41.0"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
ruff = "^0.1.6"
mypy = "^1.7.1"
isort = "^5.12.0"
types-pytz = "^2023.3.1.1"
pydantic = "^2.5.2"
taskipy = "^1.12.2"
sphinx = "^7.2.6"
sphinx-rtd-theme = "^2.0.0"
types-pyyaml = "^6.0.12.12"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ["py310"]
line-length = 120

[tool.isort]
profile = "black"

[tool.ruff]
target-version = "py311"
line-length = 120
ignore = ["E203"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"]
"**/{tests,docs,tools}/*" = ["E402"]

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
explicit_package_bases = true
exclude = ["^tests/[^/]+\\.py$", "^docs"]  # tests/**/*.pyのみ除外

[tool.taskipy.tasks]
format = {cmd = "poetry run isort . && poetry run black .", help = "Run the formatter."}  # フォーマッタを実行
lint = {cmd = "poetry run ruff . && poetry run mypy .", help = "Run the linter."}  # リンタを実行
ci = {cmd = "poetry run isort . && poetry run black . && poetry run ruff . && poetry run mypy .", help = "Run all CI tasks."}  # 全てのCIタスクを実行
test = {cmd = "poetry run pytest .", help = "Run all tests."}
dev = {cmd = "poetry run uvicorn app.main:app --reload", help = "Start the api"}
up = {cmd = "docker compose up", help = "start container"}
down = {cmd = "docker compose down", help = "stop container"}
docs = {cmd = "sphinx-apidoc --force -o docs/source/resources app/ && sphinx-build -b html docs/source/ docs/_build", help = "Generate and build documentation from sources"}
